requirements.txt
aiocontextvars==0.2.2
autopep8==1.5.2
blinker==1.4
click==7.1.2
contextvars==2.4
dominate==2.5.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Moment==0.9.0
Flask-WTF==0.14.3
immutables==0.14
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.0
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
python-dotenv==0.13.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
